fixing tests of float64

+++ ml-agents-envs/mlagents/envs/brain_conversion_utils.py
@@ -22,7 +22,7 @@ def step_result_to_brain_info(
     else:
         vec_obs = np.concatenate([step_result.obs[i] for i in vec_obs_indices], axis=1)
     vis_obs = [step_result.obs[i] for i in vis_obs_indices]
-    mask = np.ones((n_agents, np.sum(group_spec.action_shape)))
+    mask = np.ones((n_agents, np.sum(group_spec.action_shape)), dtype=np.float32)
     if step_result.action_mask is not None:
         mask = 1 - np.concatenate(step_result.action_mask, axis=1)
     if agent_id_prefix is None:
